CE-0003 Extend part info disclosure to part-type views
PartInfoCell (identity + disambiguator + (i)-tooltip) was only wired into
PartList. The relations drawer (PartTypeDetail) and the Currently Used Part
column (PartTypeList) showed the identity only, leaving two same-model parts
indistinguishable — the exact gap CE-0003 set out to fix.

Extract PartInfoCell into its own shared component and reuse it in both views.
[Claude Code - model: claude-sonnet-4-6]
